+++ frontend/dev.Dockerfile
@@ -8,8 +8,6 @@ ADD /frontend /frontend
 
 VOLUME ["/frontend"]
 
-RUN ls -l
-
 EXPOSE 3000
 EXPOSE 9876
 CMD npm install && npm run start
