+++ frontend/src/app/add-product/add-product.component.html
@@ -1,0 +1,66 @@
+<div class="add-products-container" [formGroup]="formGroup">
+    <h1>Adding products Module</h1>
+    <div class="add-products-form">
+        <mat-form-field>
+            <input matInput type="text" [formControlName]="'productName'" placeholder="Product Name">
+            <mat-hint align="end">Insert product name</mat-hint>
+        </mat-form-field>
+
+        <mat-form-field>
+            <input matInput type="number" [formControlName]="'productPrice'" placeholder="Product price">
+            <span matPrefix>PLN&nbsp;</span>
+            <mat-hint align="end">Insert cost</mat-hint>
+        </mat-form-field>
+
+        <mat-form-field>
+            <input matInput type="text" [formControlName]="'productCompanyName'" placeholder="Product Company Name">
+            <mat-hint align="end">Insert product Company name</mat-hint>
+        </mat-form-field>
+
+        <mat-form-field>
+            <input matInput type="text" [formControlName]="'productImage'" placeholder="Product Image">
+            <mat-hint align="end">Send picture of a product</mat-hint>
+        </mat-form-field>
+
+        <mat-form-field>
+            <input matInput type="text" [formControlName]="'productWeight'" placeholder="Product Weight">
+            <mat-hint align="end">Insert product weight</mat-hint>
+        </mat-form-field>
+
+        <mat-form-field>
+            <mat-select placeholder="Product Category" name="category" [formControlName]="'productCategory'">
+                <mat-option *ngFor="let category of categories" [value]="category.value">
+                    {{ category.viewValue }}
+                </mat-option>
+            </mat-select>
+            <mat-hint align="end">Choose proudct Category</mat-hint>
+        </mat-form-field>
+
+        <mat-form-field>
+            <mat-select placeholder="Product Quantitative Type" name="type" [formControlName]="'productQuantitativeType'">
+                <mat-option *ngFor="let type of qType" [value]="type.value">
+                    {{ type.viewValue }}
+                </mat-option>
+            </mat-select>
+            <mat-hint align="end">Choose product Quantitative Type</mat-hint>
+        </mat-form-field>
+
+
+        <mat-form-field>
+            <input matInput type="text" [formControlName]="'productTags'" placeholder="Product Tags">
+            <mat-hint align="end">Select product tags</mat-hint>
+        </mat-form-field>
+
+        <div class="buttons">
+            <button class="cancel-btn" mat-raised-button routerLink="/dashboard">Cancel</button>
+            <button class="cancel-btn" mat-raised-button color="secondary" routerLink="/show-products">Show All products</button>
+            <button class="add-btn" mat-raised-button color="primary" (click)="addProduct(formGroup.value)">Add</button>
+        </div>
+    </div>
+
+    <mat-card *ngIf="isErrorMsgVisible && !formGroup.valid">
+        <mat-card-content>
+            <p style="text-align: center; color: red; font-size: 16px;">Wprowadź wszystkie dane</p>
+        </mat-card-content>
+    </mat-card>
+</div>
