+++ frontend/src/app/add-product/add-product.component.scss
@@ -1,7 +1,7 @@
 :host {
     width: 100%;
 
-    .adding-products-container {
+    .add-products-container {
         width: 500px;
         margin: 0 auto;
