+++ frontend/src/app/manage-products/manage-products.component.scss
@@ -1,4 +1,4 @@
-.display-product-container {
+.manage-product-container {
     margin: 50px;
     padding: 10px;
